perf: read counters with RDPMC when possible on amd64
My primary motivation with this change is to reduce overhead of ReadOne
to make it feasible to measure counters in small regions. e.g., I would
like ignore cache misses in a routine explicitly intended to flush the
cache.

Using RDPMC with the perf_event_open API is fairly straightforward. The
primary annoyance is that the ability to use RDPMC can disappear
arbitrarily at run time. Multiplexing of many events is an obvious case
of this. That said, I saw this changing on my machine even when there
were no other PMU users that I am aware of, so it seems there are other
cases as well.

I've bumped the Go version to 1.22 for range-over-int.

goos: linux
goarch: amd64
pkg: github.com/aclements/go-perfevent/perf
cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
           │ /tmp/before.txt │            /tmp/after.txt            │
           │     sec/op      │    sec/op     vs base                │
ReadOne-12    1347.00n ± 15%   80.21n ± 50%  -94.05% (p=0.000 n=10)

(It is ~25ns/op without a serializing instruction. It would be
interesting to know if we could elide that in favor of something cheaper
like MFENCE depending on the event.)